Move base attribute to core
