+++ MvvmCross/Core/Core/MvvmCross.Core.csproj
@@ -133,6 +133,8 @@
     <Compile Include="Navigation\MvxNavigationCache.cs" />
     <Compile Include="Annotations\JetbrainsAnnotations.cs" />
     <Compile Include="ViewModels\MvxNavigationServiceAppStart.cs" />
+    <Compile Include="Views\MvxBasePresentationAttribute.cs" />
+    <Compile Include="Views\IMvxOverridePresentationAttribute.cs" />
   </ItemGroup>
   <ItemGroup />
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
