+++ MvvmCross/Core/Core/Views/MvxBasePresentationAttribute.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 
-namespace MvvmCross.Droid.Views.Attributes
+namespace MvvmCross.Core.Views
 {
     [AttributeUsage(AttributeTargets.Class)]
     public abstract class MvxBasePresentationAttribute : Attribute
     {
         /// <summary>
-        /// That shall be used only if you are using non generic fragments.
+        /// That shall be used only if you are using non generic views.
         /// </summary>
         public Type ViewModelType { get; set; }
