--- MvvmCross/Droid/Droid/Views/Attributes/IMvxOverridePresentationAttribute.cs
@@ -1,7 +1,0 @@
-﻿namespace MvvmCross.Droid.Views.Attributes
-{
-    public interface IMvxOverridePresentationAttribute
-    {
-        MvxBasePresentationAttribute PresentationAttribute();
-    }
-}
