+++ MvvmCross/Droid/Droid/Views/Attributes/MvxFragmentAttribute.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Android.Views;
+using MvvmCross.Core.Views;
 
 namespace MvvmCross.Droid.Views.Attributes
 {
