--- MvvmCross/Mac/Mac/Views/Presenters/Attributes/IMvxOverridePresentationAttribute.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace MvvmCross.Mac.Views.Presenters.Attributes
-{
-    public interface IMvxOverridePresentationAttribute
-    {
-        MvxBasePresentationAttribute PresentationAttribute();
-    }
-}
