--- MvvmCross/Mac/Mac/Views/Presenters/Attributes/MvxBasePresentationAttribute.cs
@@ -1,8 +1,0 @@
-﻿using System;
-namespace MvvmCross.Mac.Views.Presenters.Attributes
-{
-    [AttributeUsage(AttributeTargets.Class)]
-    public abstract class MvxBasePresentationAttribute : Attribute
-    {
-    }
-}
