+++ MvvmCross/Mac/Mac/Views/Presenters/Attributes/MvxSheetPresentationAttribute.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using MvvmCross.Core.Views;
+
 namespace MvvmCross.Mac.Views.Presenters.Attributes
 {
     public class MvxSheetPresentationAttribute : MvxBasePresentationAttribute
