+++ MvvmCross/Mac/Mac/Views/Presenters/Attributes/MvxWindowPresentationAttribute.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using AppKit;
+using MvvmCross.Core.Views;
 
 namespace MvvmCross.Mac.Views.Presenters.Attributes
 {
