--- MvvmCross/iOS/iOS/Views/Presenters/Attributes/MvxBasePresentationAttribute.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace MvvmCross.iOS.Views.Presenters.Attributes
-{
-    [AttributeUsage(AttributeTargets.Class)]
-    public abstract class MvxBasePresentationAttribute : Attribute
-    {
-    }
-}
