+++ MvvmCross/iOS/iOS/Views/Presenters/Attributes/MvxChildPresentationAttribute.cs
@@ -1,4 +1,6 @@
-﻿namespace MvvmCross.iOS.Views.Presenters.Attributes
+﻿using MvvmCross.Core.Views;
+
+namespace MvvmCross.iOS.Views.Presenters.Attributes
 {
     public class MvxChildPresentationAttribute : MvxBasePresentationAttribute
     {
