+++ MvvmCross/iOS/iOS/Views/Presenters/Attributes/MvxRootPresentationAttribute.cs
@@ -1,4 +1,6 @@
-﻿namespace MvvmCross.iOS.Views.Presenters.Attributes
+﻿using MvvmCross.Core.Views;
+
+namespace MvvmCross.iOS.Views.Presenters.Attributes
 {
     public class MvxRootPresentationAttribute : MvxBasePresentationAttribute
     {
